re-implement /admin/sign-out with logout logic
okay so the logout() endpoint will destroy the local cookies but NOT
log out of OKTA itself, so clicking the Admin link will just open the
page right back up without re-auth, which might be weird. The other
commented alternative ends the OKTA session to force a reauth, i'm not
sure which will be better.